Removed override on background color in dialog box
Having browser in dark mode sets text color to white in the dialog box. The css style for .ido-module-dialog sets background color to white, without setting text color to black. This results in white text on white background when the browser is in dark mode.

An alternative fix is to set text color to black at the same time as background color is set to white.